MSEDataAnalising/requirements.txt
pip~=23.2.1
attrs~=24.3.0
argparse~=1.4.0
wheel~=0.45.1
cryptography~=44.0.0
pillow~=11.0.0
utils~=1.0.2
automium_web~=0.1.1
filelock~=3.16.1
Django~=5.1.4
asgiref~=3.8.1
sqlparse~=0.5.3
psycopg2~=2.9.10
colorama~=0.4.6
PyYAML~=6.0.2
Jinja2~=3.1.5
numpy~=2.0.2
requests~=2.32.3
serializers~=0.2.4
typing_extensions~=4.12.2
views~=0.3
djangorestframework~=3.15.2
services~=0.1.1
torch~=2.5.1
transformers~=4.47.1
nlp~=0.4.0
scikit-learn~=1.6.0
tqdm~=4.67.1
beautifulsoup4~=4.12.3
pdfplumber~=0.11.5
pandas~=2.2.3
joblib~=1.4.2
tensorflow~=2.18.0
keras~=3.8.0
ta~=0.11.0
statsmodels~=0.14.4
selenium~=4.27.1
Flask~=3.1.0
matplotlib~=3.10.0